+++ src/main/resources/ci3-stubs/CI_Controller.php
@@ -24,6 +24,7 @@
  * @property CI_Javascript $jquery
  * @property CI_Calendar $calendar
  * @property CI_Upload $upload
+ * @property CI_Lang $lang
  */
 class CI_Controller {
 }